NOISSUE: refactoring
