+++ src/components/Wizard/Wizard.vue
@@ -13,7 +13,7 @@ import { useSlideBuilder } from '../../composables/useSlideBuilder';
 import { RouteNames } from '../../routes';
 import type { Slide } from '../../types/slide-schema';
 import FolderPicker from '../FolderPicker/FolderPicker.vue';
-import { ipcRenderer } from 'electron';
+import { loadBackupImages, pickRandomNumbers } from './Wizard.util';
 
 const presentation = usePresentation();
 const { push } = useRouter();
@@ -40,23 +40,6 @@ watch(duration, (value: number) => {
     timePerTick: value * 1000,
   };
 });
-
-function* range(toExcl: number): Generator<number> {
-  for (let i = 0; i < toExcl; i++) yield i;
-}
-
-function pickRandomNumbers(howMany: number, toExcl: number): number[] {
-  const available: number[] = Array.from(range(toExcl));
-  if (howMany >= toExcl) return available;
-
-  const picked: number[] = [];
-  while (picked.length < howMany) {
-    const pickedIndex = Math.ceil(Math.random() * available.length);
-    const [pickedValue] = available.splice(pickedIndex, 1);
-    picked.push(pickedValue);
-  }
-  return picked;
-}
 
 interface GenerationOptions {
   topic: string;
@@ -85,14 +68,6 @@ async function generateSlide({ topic, slideNo, withImage, withText }: Generation
   return slide;
 }
 
-async function loadBackupImages(): Promise<string[]> {
-  console.debug({ localImagePath: localImagePath.value });
-  const howMany = length.value - images.value;
-  if (!howMany) return [];
-
-  return await ipcRenderer.invoke('pick-random-images', localImagePath.value, howMany);
-}
-
 const generate = () => {
   tasksDone.value = 0;
   tasksTotal.value = 0;
@@ -100,7 +75,9 @@ const generate = () => {
   const slidesWithAiImage = new Set<number>(Array.from(pickRandomNumbers(images.value, length.value)));
 
   op(async () => {
-    const backupImages = localImagePath.value ? await loadBackupImages() : [];
+    const backupImages = localImagePath.value
+      ? await loadBackupImages(localImagePath.value, length.value - images.value)
+      : [];
 
     if (!presentation.topic) {
       presentation.topic = await findTopic();
@@ -120,7 +97,6 @@ const generate = () => {
 
         if (!withAiImage && backupImages.length) {
           const [localImage] = backupImages.splice(0, 1);
-          console.debug('Using local image', { slide: i, localImage });
           slide.image = { base64: localImage };
         }
